+++ native/core/src/common/bit.rs
@@ -141,17 +141,6 @@ pub fn read_u32(src: &[u8]) -> u32 {
     unsafe { in_ptr.read_unaligned() }
 }
 
-/// Similar to the `read_num_bytes` but read nums from bytes in big-endian order
-/// This is used to read bytes from Java's OutputStream which writes bytes in big-endian
-macro_rules! read_num_be_bytes {
-    ($ty:ty, $size:expr, $src:expr) => {{
-        debug_assert!($size <= $src.len());
-        let mut buffer = <$ty as $crate::common::bit::FromBytes>::Buffer::default();
-        buffer.as_mut()[..$size].copy_from_slice(&$src[..$size]);
-        <$ty>::from_be_bytes(buffer)
-    }};
-}
-
 #[inline]
 pub fn memcpy(source: &[u8], target: &mut [u8]) {
     debug_assert!(target.len() >= source.len(), "Copying from source to target is not possible. Source has {} bytes but target has {} bytes", source.len(), target.len());
