+++ native/core/src/execution/expressions/mod.rs
@@ -17,8 +17,6 @@
 
 //! Native DataFusion expressions
 
-pub mod bloom_filter_agg;
-pub mod bloom_filter_might_contain;
 pub mod subquery;
 
 pub use datafusion_comet_spark_expr::EvalMode;
