+++ native/core/src/execution/planner.rs
@@ -24,7 +24,6 @@ use crate::{
     errors::ExpressionError,
     execution::{
         expressions::{
-            bloom_filter_agg::BloomFilterAgg, bloom_filter_might_contain::BloomFilterMightContain,
             subquery::Subquery,
         },
         operators::{CopyExec, ExecutionError, ExpandExec, ScanExec},
@@ -707,17 +706,6 @@ impl PhysicalPlanner {
                 let id = expr.id;
                 let data_type = to_arrow_datatype(expr.datatype.as_ref().unwrap());
                 Ok(Arc::new(Subquery::new(self.exec_context_id, id, data_type)))
-            }
-            ExprStruct::BloomFilterMightContain(expr) => {
-                let bloom_filter_expr = self.create_expr(
-                    expr.bloom_filter.as_ref().unwrap(),
-                    Arc::clone(&input_schema),
-                )?;
-                let value_expr = self.create_expr(expr.value.as_ref().unwrap(), input_schema)?;
-                Ok(Arc::new(BloomFilterMightContain::try_new(
-                    bloom_filter_expr,
-                    value_expr,
-                )?))
             }
             ExprStruct::CreateNamedStruct(expr) => {
                 let values = expr
@@ -1982,20 +1970,6 @@ impl PhysicalPlanner {
                     expr.null_on_divide_by_zero,
                 ));
                 Self::create_aggr_func_expr("correlation", schema, vec![child1, child2], func)
-            }
-            AggExprStruct::BloomFilterAgg(expr) => {
-                let child = self.create_expr(expr.child.as_ref().unwrap(), Arc::clone(&schema))?;
-                let num_items =
-                    self.create_expr(expr.num_items.as_ref().unwrap(), Arc::clone(&schema))?;
-                let num_bits =
-                    self.create_expr(expr.num_bits.as_ref().unwrap(), Arc::clone(&schema))?;
-                let datatype = to_arrow_datatype(expr.datatype.as_ref().unwrap());
-                let func = AggregateUDF::new_from_impl(BloomFilterAgg::new(
-                    Arc::clone(&num_items),
-                    Arc::clone(&num_bits),
-                    datatype,
-                ));
-                Self::create_aggr_func_expr("bloom_filter_agg", schema, vec![child], func)
             }
         }
     }
